dags/pyproject.toml
[project]
name = "dags"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "load-dotenv (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
